requirements.txt
flask
azure-storage
flask-cors